+++ server/service/LlmPromptBuilderService.php
@@ -463,7 +463,8 @@ class LlmPromptBuilderService extends BaseLlmService
             $model_name,
             $temperature,
             $max_tokens,
-            $section_id ?: null
+            $section_id ?: null,
+            LLM_CONV_SOURCE_BUILDER
         );
     }
 }