+++ server/service/LlmPromptPlaygroundService.php
@@ -889,7 +889,8 @@ class LlmPromptPlaygroundService extends BaseLlmService
             $model_name,
             $temperature,
             $max_tokens,
-            $section_id ?: null
+            $section_id ?: null,
+            LLM_CONV_SOURCE_PLAYGROUND
         );
     }
 }